% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_method.R
\name{s7_method}
\alias{s7_method}
\alias{s7_method,integer-method}
\title{An S7 generic}
\usage{
s7_method(x, ...)

## S7 method for class <integer>
s7_method(x, ...)
}
\arguments{
\item{x, y}{A parameter}
}
\description{
An S7 generic
}
