% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_multi.R
\name{s7_multi}
\alias{s7_multi}
\alias{s7_multi,integer,character-method}
\title{An S7 multi-dispatch generic}
\usage{
s7_multi(x, y, ...)

## S7 method for classes <integer>, <character>
s7_multi(x, y, ...)
}
\arguments{
\item{x, y}{A parameter}
}
\description{
An S7 multi-dispatch generic
}
